Add support for <link rel=preload>
This adds support for <link rel=preload> when the server is an HTTP1
server. This is through a header like:

```
Link: </dist/bundles/index.js>; rel=preload; as=script;
```

Closes #4
