+++ lib/index.js
@@ -23,26 +23,42 @@ function create({manifest: manifestPath, cwd = process.cwd()}) {
 		assets = makeIterable(assets);
 
 		return function(req, res, next = noop){
-			if(typeof req.push !== "function") {
-				return next();
-			}
+			if(typeof res.push !== "function") {
+				for(let [pth, asset] of assets) {
+					let serverPath = `/${pth}`;
+					let header;
 
-			for(let [pth, asset] of assets) {
-				let readStream = fs.createReadStream(
-					path.join(cwd, pth)
-				);
-				let serverPath = `/${pth}`;
+					switch(asset.type) {
+						case "script":
+							header = `<${serverPath}>; rel=preload; as=script;`;
+							break;
+						case "style":
+							header = `<${serverPath}>; rel=preload; as=style;`;
+							break;
+						default:
+							return next();
+					}
 
-				switch(asset.type) {
-					case "script":
-						pushScript(req, res, readStream, serverPath);
-						break;
-					case "style":
-						pushStyle(req, res, readStream, serverPath);
-						break;
-					default:
-						// What do? not push for now
-						return next();
+					res.setHeader("Link", header);
+				}
+			} else {
+				for(let [pth, asset] of assets) {
+					let readStream = fs.createReadStream(
+						path.join(cwd, pth)
+					);
+					let serverPath = `/${pth}`;
+
+					switch(asset.type) {
+						case "script":
+							pushScript(req, res, readStream, serverPath);
+							break;
+						case "style":
+							pushStyle(req, res, readStream, serverPath);
+							break;
+						default:
+							// What do? not push for now
+							return next();
+					}
 				}
 			}
